// rooms-webapp/src/generated/rooms-api.ts
export interface components {
    schemas: {
        Room: {
            id: string;
            name: string;
            capacity: number;
            location: string;
        };
        NewBooking: {
            roomId: string;
            /** Format: date */
            date: string;
            /** Format: time */
            startTime: string;
            /** Format: time */
            endTime: string;
            attendeeCount: number;
        };
        Booking: {
            id: string;
            roomId: string;
            roomName: string;
            /** Format: date */
            date: string;
            /** Format: time */
            startTime: string;
            /** Format: time */
            endTime: string;
            attendeeCount: number;
            /** Format: date-time */
            createdAt: string;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}